requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
httplib2==0.22.0
idna==3.6
multidict==6.0.5
oauthlib==3.2.2
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyTelegramBotAPI==4.16.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
telebot==0.0.5
tqdm==4.66.2
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
Flask
python-telegram-bot-calendar
watchdog